A new PMML parser
Full recursive-descent style parsing.  MUCH easier to track the control flows
than a SAX-style parser, and does not require the entire doc like a DOM-style.
More tightly tied to the XML structure, but in theory PMML is a multi-vendor
standard and fairly stable.

Like a good R-D parser, uses a separate function for parsing each XML element.
Each function expects to be at a particular parse-point (generally after the
openning '<' and before the tag is parsed), and always leaves the parse just
after the close-tag '>'.  The semantic interpretation is then interleaved with
the parsing, with higher levels passing down needed info to lower element
levels, and lower levels directly returning results to the higher levels.