Toss out columns which are >20% NA's.
Helps most algo's which don't handle mostly-NA-cols by default.
Also, in pass2 parsing if the col is declared a STRINGCOL then
do not parse it as a number (then fail back to string parsing).
Just start out with string parsing.  In the matching tests, increase
the dataset sizes to prevent NA'ing whole columns.