Gut Paxos.  New simpler consensus building algorithm
Detects and builds a "cloud" - a cooperating group of nodes, with mutual
knowledge of each other.  Basically tracks all the nodes that *this* node
has ever heard of, and when *all* of the other nodes have all heard of each
other, declares the situation as "commonKnowledge", and a Cloud.
This algorithm differs from Paxos in a number of obvious ways:
 * - it is not robust against failing nodes
 * - it requires true global consensus (a Quorum of All)
 * - it is vastly simpler than Paxos