Fix bug parser reduce
In the rare case that a DTask parse fits in a UDP packet, and splits on
multiple nodes, and the first attempt to send drops the packet, then the
local node makes progress filling in fields of _nrows.  On the UDP retry,
a non-zero nrows is shipped to the remote, who faithfully fills in all the
other fields, then ships back an nrows with bits set in some entries that
where set by the originating node.  In the reduction phase, these where
getting added - effectively doubling the nrows count.  Switch to OR'ing
since the bits have to be either zero or equal.
Also fix a name-change from _ab to _abs, and add some asserts.