Bugfix in GLMGrid.
KV Store changed and now stores deserialized objects so objects obtained from KV  are now shared and not private.
GLMGrid contains sort of it's array of keys (to order models according to their goodness), which conflicted with the update as new models were computed.

Changed it so that it now sorts cloned array of keys.