Major change: track completed tasks For All Time
Never ever confuse a very late UDP packet with fresh work - by tracking
task#'s forever.  To prevent leaking task objects, completed tasks are
collapsed to a single integer ("all tasks before this number are completed").
This means that every task# has to be sent from client to server; the client
cannot skip any numbers (or the server thinks it's got a not-yet-completed
task outstanding).  To ease life, the new invariant is that the combination
of (h2onode,task#) is unique across the cluster; task#'s are track independently
between any 2 nodes (and even in each direction; since we're peer-to-peer
each node is both a client and a server to each other node).
The RPC.TASKS lists is now split and run independently in each H2ONode object.
The unique-task# is now split and run in each H2ONode.
Same for a completed task# counter.
Attempts to initiate a task# that is in the range of the completed task#
counter will act "as if" we're looking at a very old completed task.