YADLF - Yet Another Distributed Logging Framework
Only kinda-sorta: if "-log" is passed on the cmd-line, hook System.out and
System.err to add headers, and also write them to ice_root/h2o.log, and also
append them to the internal system key "Log".  Add a debug-link to view the
log - in memory it only keeps the last 1000 log msgs across the cluster.  They
are stored in the order recieved (which may not match the time stamps).  The
local log files are *local* log messages only.