Update of the Multifile parse for the fluid vecs. Mostly updated dealing with enums.
In case of multi file parse, the result chunks will generally not be at the same nodes as the input chunks, which causes problems for enum handling as we need to update the enums in the end to match the global encoding given by Vector's domain. To be able to do that, we need to get the node-local-Enum which was used for the first round of enum encoding for each chunk. That translates to knowing on which node was each of the output chunks produced. The parser task now remembers the mapping from target chunks to nodes. We than need to fetch the local enums from all the nodes and pass all this info to the enum update task to get correct mapping.