Break out the boolean response-variable case
The one-vs-all classifier strategy makes K boolean predictors; boolean
predictors can take shortcuts - the responses are only ever 0 or 1, and
the squares of 0 or 1 are also 0 or 1 respectively.  Means the
sum-of-squares is also the sum, and a bunch of math simplifies.
Also move the AtomicXXXArray updaters into the Utils directory.
Have DRealHistogram and DBinomHistogram share DHistogram, which carries
the bulk of the code.  Got lots of virtual calls (might have to be
hand-inlined to dodge the v-call costs).  But uses a bunch less memory,
and can use a bunch-less-more with more work.