Fix of HEX-1308: vector alignment in confusion matrix.
For CM construction a user can select two vectors from
two different sources. Hence, vectors do not need to be in the
same vector group.
This fix check vector group membership and if the vector are in
different group, then one of the vectors is aligned according to
the distribution of another vector. Then CM is constructed.
At the end, newly created vector is deleted.