Make the python multi-JVM NeuralNet Mnist test work again, and similarly to the single-VM case, test for accuracy in validation set prediction, both via model progress page and via separate scoring call.
For now, only use the multi-threaded single-node mode, even on 3 JVMs.  This
means that only one JVM actually affects the results,  Due to chunking, this
will lead to worse accuracy than if a truly distributed (synchronized) mode was
used. But at least this test tests that the code will run, until the
'MapReduce' mode is implemented, and data and weight matrices from all machines
are taken into account.