Remove all matching H2O keys before importing.
Change "import folder" to "sync folder".  Remove all keys with
file-prefix names matching the folder, before importing.  This will
force a re-load of data in the same directory - so if you've removed a
junk file from a directory & re-import, you'll have keys that exactly
match the directory.  Ditto for editing a file & re-importing.