On the way to fix PUB-65, HEX-1581, PUB-76 - tricky +1 error.
For GBM/DRF - if validation dataset is not specified
then on-the-fly data are used for scoring.

BUT, if validation dataset is specified model is used for scoring.
BUT it was updated after scoring. So all MSE reports were shifted since
there were two MS errors reports based on model containing 0-trees.

No, model is updated by appending a new tree and then updated again
by updating MSE, and CM records. Not fancy.