taking a chance here. I don't think this is a final answer, but will be interesting to see what Irene thinks for R compare. I did some compares and I think it matches our R desires better. Basically return the estimated sampleQuantile always for Summary2. Don't use the new sampleQuantile that was introduced for Summary2.
And get rid of the 0.5*.. addition, which I think gets you the estimated midpoint for the bin implied by the threshold. For R, they want the 0.33/0.5/0.66 best estimates. Since our thresholds define those as bin edges (starts) ..the estimated midpoint, is not the best estimate for R.

We have the problem that for emulation of other quantile bin reports, we probably want the estimates for centers of a 0.5 "bin". it all depends on whether you think a 0.5 bin starts at 0.5, or is centered at 0.5. We make it start at 0.5 but return an estimate for 0.5 + half bin.

I put // kbn in the two places changed for easy undo. Only Summary2

here's where I think the half bin thing was
      // kbn. change so we get estimate at the threshold, not midpoint of bin.
      // I think this matches what R wants for a threshold. But probably doesn't match
      // what a binned quantile wants. (should we produce both?)
      qtiles[j] = _mins[0] + k*_binsz;
      // was, I think for midpoint estimation of the bin implied by the thresholds
      // qtiles[j] = _mins[0] + k*_binsz + ((_binsz > 1)?0.5*_binsz:0);