Added asyncExec call to MRTask2 intended to be used for non-blocking task invocation (either using continuation/callback style or do not wait for the results at all).
The functionality is basically the same as of dfork call, except the invoked task will run at the same priority level as the calling thread or min_priority if called outside of FJ framework (whereas dfork automatically increases the priority level to avoid deadlock).

Adapted GLM2 (and Parse2) to use this call instead of dfork so that every GLM2 iteration runs at the same priority level (GLM2 launches iterations from FJ thread but uses continuation style callback and not blocking, so there is no risk of deadlock).