Generic redirector for Inspect key calls.
Fix for PUB-590.

The fix introduces a new request class which
redirects based on key type (needs to fetch the value) to
a page serving value view.

Now the redirect is generic for all Models. If the value
is assignable to Model.class then the class redirects to
<model_name>View.html?source= page.

If the value is a frame, then the page redirects to Inspect2.